+++ app/src/main/java/fr/jaetan/jmedia/controllers/BookController.kt
@@ -3,6 +3,7 @@ package fr.jaetan.jmedia.controllers
 import androidx.compose.runtime.mutableStateListOf
 import fr.jaetan.jmedia.core.networking.BookApi
 import fr.jaetan.jmedia.core.realm.entities.toBooks
+import fr.jaetan.jmedia.core.realm.repositories.BookRepository
 import fr.jaetan.jmedia.models.works.Book
 import fr.jaetan.jmedia.models.works.takeWhereEqualTo
 import fr.jaetan.jmedia.models.works.toBdd
@@ -12,6 +13,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
 class BookController : IWorkController<Book>() {
+    private val repository by lazy { BookRepository(MainViewModel.realm) }
     override val fetchedWorks = mutableStateListOf<Book>()
     override var localWorks = mutableStateListOf<Book>()
 
@@ -25,7 +27,7 @@ class BookController : IWorkController<Book>() {
 
     override suspend fun initializeFlow() {
         CoroutineScope(Dispatchers.IO).launch {
-            MainViewModel.bookRepository.all.collect {
+            repository.all.collect {
                 localWorks.clear()
                 localWorks.addAll(it.list.toBooks())
                 setLibraryValues()
@@ -42,11 +44,15 @@ class BookController : IWorkController<Book>() {
     override suspend fun libraryHandler(work: Book) {
         if (work.isInLibrary) {
             localWorks.takeWhereEqualTo(work)?.let {
-                MainViewModel.bookRepository.remove(it.toBdd())
+                repository.remove(it.toBdd())
             }
             return
         }
 
-        MainViewModel.bookRepository.add(work.toBdd())
+        repository.add(work.toBdd())
+    }
+
+    override suspend fun removeAll() {
+        repository.removeAll()
     }
 }