+++ app/src/main/java/fr/jaetan/jmedia/controllers/MovieController.kt
@@ -3,6 +3,7 @@ package fr.jaetan.jmedia.controllers
 import androidx.compose.runtime.mutableStateListOf
 import fr.jaetan.jmedia.core.networking.MovieApi
 import fr.jaetan.jmedia.core.realm.entities.toMovies
+import fr.jaetan.jmedia.core.realm.repositories.MovieRepository
 import fr.jaetan.jmedia.models.works.Movie
 import fr.jaetan.jmedia.models.works.takeWhereEqualTo
 import fr.jaetan.jmedia.models.works.toBdd
@@ -12,6 +13,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
 class MovieController : IWorkController<Movie>() {
+    private val repository by lazy { MovieRepository(MainViewModel.realm) }
     override val fetchedWorks = mutableStateListOf<Movie>()
     override var localWorks = mutableStateListOf<Movie>()
 
@@ -25,7 +27,7 @@ class MovieController : IWorkController<Movie>() {
 
     override suspend fun initializeFlow() {
         CoroutineScope(Dispatchers.IO).launch {
-            MainViewModel.movieRepository.all.collect {
+            repository.all.collect {
                 localWorks.clear()
                 localWorks.addAll(it.list.toMovies())
                 setLibraryValues()
@@ -42,7 +44,7 @@ class MovieController : IWorkController<Movie>() {
     override suspend fun libraryHandler(work: Movie) {
         if (work.isInLibrary) {
             localWorks.takeWhereEqualTo(work)?.let {
-                MainViewModel.movieRepository.remove(it.toBdd())
+                repository.remove(it.toBdd())
             }
             return
         }
@@ -59,6 +61,10 @@ class MovieController : IWorkController<Movie>() {
             else it
         }
 
-        MainViewModel.movieRepository.add(movie.toBdd())
+        repository.add(movie.toBdd())
+    }
+
+    override suspend fun removeAll() {
+        repository.removeAll()
     }
 }