+++ app/src/main/java/fr/jaetan/jmedia/models/BuildType.kt
@@ -5,15 +5,14 @@ import fr.jaetan.jmedia.BuildConfig
 enum class BuildType {
     Release,
     Staging,
+    Demo,
     Debug;
-
-    val isInRelease: Boolean
-        get() = this == Release
 
     companion object {
         fun get(): BuildType = when (BuildConfig.BUILD_TYPE) {
             "release" -> Release
             "staging" -> Staging
+            "demo" -> Demo
             else -> Debug
         }
     }