+++ app/src/main/java/fr/jaetan/jmedia/services/GlobalSettings.kt
@@ -7,11 +7,15 @@ import fr.jaetan.jmedia.models.BuildType
 object GlobalSettings {
     private val buildType = BuildType.get()
 
-    const val versionName = BuildConfig.VERSION_NAME
     const val versionCode = BuildConfig.VERSION_CODE
+    const val fullVersionName = BuildConfig.VERSION_NAME
+    val versionName = when {
+        fullVersionName.contains("-") -> fullVersionName.split("-").first()
+        else -> fullVersionName
+    }
 
-    val isInRelease: Boolean
-        get() = buildType.isInRelease
+    val isInDemo: Boolean
+        get() = buildType == BuildType.Demo
 
     enum class ApiKeys(val key: String) {
         Github(BuildConfig.GITHUB_API_KEY),